backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
websockets==12.0
torch==2.1.1
numpy==1.26.2
nltk==3.8.1
aiosqlite==0.19.0
bcrypt==4.1.2
